Support zero argument defs
