+++ core/src/main/scala/org/bykn/bosatsu/Statement.scala
@@ -1,6 +1,6 @@
 package org.bykn.bosatsu
 
-import Parser.{ Combinators, lowerIdent, upperIdent, maybeSpace, spaces }
+import Parser.{ Combinators, Indy, lowerIdent, upperIdent, maybeSpace, spaces }
 import cats.Functor
 import cats.data.{ NonEmptyList, State }
 import cats.implicits._
@@ -8,61 +8,8 @@ import fastparse.all._
 import org.typelevel.paiges.{ Doc, Document }
 
 import org.bykn.fastparse_cats.StringInstances._
-import Parser.Indy
 
 import Indy.IndyMethods
-
-case class DefStatement[T](
-    name: String,
-    args: NonEmptyList[(String, Option[TypeRef])],
-    retType: Option[TypeRef], result: T) {
-
-
-  /**
-   * This ignores the name completely and just returns the lambda expression here
-   *
-   * This could be a traversal: TypeRef => F[rankn.Type] for an Applicative F[_]
-   */
-  def toLambdaExpr[A](resultExpr: Expr[A], tag: A)(trFn: TypeRef => rankn.Type): Expr[A] = {
-    val unTypedBody = resultExpr
-    val bodyExp =
-      retType.fold(unTypedBody) { t =>
-        Expr.Annotation(unTypedBody, trFn(t), tag)
-      }
-    val deepFunctor = Functor[NonEmptyList].compose[(String, ?)].compose[Option]
-    Expr.buildLambda(deepFunctor.map(args)(trFn), bodyExp, tag)
-  }
-}
-
-object DefStatement {
-  private[this] val defDoc = Doc.text("def ")
-
-  implicit def document[T: Document]: Document[DefStatement[T]] =
-    Document.instance[DefStatement[T]] { defs =>
-      import defs._
-      val res = retType.fold(Doc.empty) { t => Doc.text(" -> ") + t.toDoc }
-      val line0 = defDoc + Doc.text(name) + Doc.char('(') +
-        Doc.intercalate(Doc.text(", "), args.toList.map(TypeRef.argDoc _)) +
-        Doc.char(')') + res + Doc.text(":")
-      line0 + Document[T].document(result)
-    }
-
-    /**
-     * The resultTParser should parse some indentation any newlines
-     */
-    def parser[T](resultTParser: P[T]): P[DefStatement[T]] = {
-      val args = argParser.nonEmptyList
-      val result = P(maybeSpace ~ "->" ~/ maybeSpace ~ TypeRef.parser).?
-      P("def" ~ spaces ~/ lowerIdent ~ "(" ~ maybeSpace ~ args ~ maybeSpace ~ ")" ~
-        result ~ maybeSpace ~ ":" ~/ resultTParser)
-        .map {
-          case (name, args, resType, res) => DefStatement(name, args, resType, res)
-        }
-    }
-
-    val argParser: P[(String, Option[TypeRef])] =
-      P(lowerIdent ~ (":" ~/ maybeSpace ~ TypeRef.parser).?)
-}
 
 case class BindingStatement[B, T](name: B, value: Declaration, in: T)
