+++ core/src/test/scala/org/bykn/bosatsu/ParserTest.scala
@@ -331,7 +331,7 @@ foo"""
     parseTestAll(
       Declaration.parser(""),
       defWithComment,
-      Declaration.DefFn(DefStatement("foo", NonEmptyList.of(("a", None)), None,
+      Declaration.DefFn(DefStatement("foo", List(("a", None)), None,
         (OptIndent.paddedIndented(1, 2, Declaration.Comment(CommentStatement(NonEmptyList.of(" comment here"),
           Padding(0, Declaration.Var("a"))))),
          Padding(0, Declaration.Var("foo"))))))
